# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AIAISummaryRequest:
    properties:
      content:
        description: 文章内容
        type: string
    required:
    - content
    type: object
  dto.AIAISummaryResponse:
    properties:
      summary:
        description: 生成的AI摘要
        type: string
    type: object
  dto.AISummaryRequest:
    properties:
      content:
        description: 文章内容
        type: string
    required:
    - content
    type: object
  dto.AISummaryResponse:
    properties:
      summary:
        description: 生成的摘要
        type: string
    type: object
  dto.AITitleRequest:
    properties:
      content:
        description: 文章内容
        type: string
    required:
    - content
    type: object
  dto.AITitleResponse:
    properties:
      title:
        description: 生成的标题
        type: string
    type: object
...
  dto.ArticleContribution:
    properties:
      count:
        description: 文章发布数量
        type: integer
      date:
        description: 日期 YYYY-MM-DD
        type: string
    type: object
...
  dto.CategoryStats:
    properties:
      count:
        description: 文章数量
        type: integer
      name:
        description: 分类名称
        type: string
    type: object
...
  dto.CollectRequest:
    properties:
      article_id:
        description: 文章ID（可选）
        type: integer
      duration:
        description: 停留时长（秒）
        type: integer
      event_data:
        additionalProperties: true
        description: 事件数据
        type: object
      event_name:
        description: 事件名称
        type: string
      hostname:
        description: 主机名
        type: string
      language:
        description: 浏览器语言
        type: string
      referrer:
        description: 来源页面
        type: string
      screen:
        description: 屏幕分辨率
        type: string
      timestamp:
        description: 时间戳
        type: integer
      title:
        description: 页面标题
        type: string
      type:
        description: '类型: pageview/event/duration'
        type: string
      url:
        description: 访问URL
        type: string
    type: object
...
  dto.DashboardStats:
    properties:
      comments_growth:
        description: 评论数增长率
        type: number
      today_comments:
        description: 今日评论数
        type: integer
      today_users:
        description: 今日注册用户
        type: integer
      today_views:
        description: 今日统计
        type: integer
      today_visitors:
        description: 今日访客量
        type: integer
      total_articles:
        description: 基础统计
        type: integer
      total_comments:
        description: 总评论数
        type: integer
      total_friends:
        description: 友链数
        type: integer
      total_moments:
        description: 动态数
        type: integer
      total_users:
        description: 总用户数
        type: integer
      total_views:
        description: 总浏览量(PV)
        type: integer
      total_visitors:
        description: 总访客量(UV)
        type: integer
      users_growth:
        description: 用户增长率
        type: number
      views_growth:
        description: 趋势对比（较昨日）
        type: number
      visitors_growth:
        description: 访客量增长率
        type: number
    type: object
...
  dto.TagStats:
    properties:
      count:
        description: 文章数量
        type: integer
      name:
        description: 标签名称
        type: string
    type: object
  dto.TrendData:
    properties:
      date:
        description: 日期
        type: string
      pv_count:
        description: 浏览量
        type: integer
      uv_count:
        description: 访客量
        type: integer
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact:
    email: talen2004@163.com
    name: M.Talen
  description: 一个基于 Go 语言的现代化博客后端服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Flec-Server
  version: v1
paths:
  /admin/ai/ai-summary:
    post:
      consumes:
      - application/json
      description: 基于文章内容生成AI摘要（150-200字，旁观者角度）
      parameters:
      - description: 文章内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AIAISummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AIAISummaryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 生成AI摘要
      tags:
      - AI功能
  /admin/ai/summary:
    post:
      consumes:
      - application/json
      description: 基于文章内容自动生成摘要（50-100字，创作者角度）
      parameters:
      - description: 文章内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AISummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AISummaryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 生成文章摘要
      tags:
      - AI功能
  /admin/ai/title:
    post:
      consumes:
      - application/json
      description: 根据内容生成多个标题建议
      parameters:
      - description: 文章内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AITitleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AITitleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 生成标题建议
      tags:
      - AI功能
...
  /admin/stats/category:
    get:
      consumes:
      - application/json
      description: 每个分类的文章数量统计
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CategoryStats'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 分类统计
      tags:
      - 统计管理
  /admin/stats/contribution:
    get:
      consumes:
      - application/json
      description: 获取文章发布数据，支持按年份或月份查询
      parameters:
      - description: 年份（可选）
        example: 2025
        in: query
        name: year
        type: integer
      - description: 月份 1-12（可选）
        example: 11
        in: query
        name: month
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ArticleContribution'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 文章贡献图
      tags:
      - 统计管理
  /admin/stats/dashboard:
    get:
      consumes:
      - application/json
      description: 获取基础统计、今日数据和趋势对比
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DashboardStats'
              type: object
      security:
      - BearerAuth: []
      summary: 仪表盘统计
      tags:
      - 统计管理
  /admin/stats/tag:
    get:
      consumes:
      - application/json
      description: 每个标签的文章数量统计
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.TagStats'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 标签统计
      tags:
      - 统计管理
  /admin/stats/trend:
    get:
      consumes:
      - application/json
      description: 指定时间段的访问趋势，支持按天/周/月聚合
      parameters:
      - description: 开始日期 YYYY-MM-DD
        example: "2025-09-01"
        in: query
        name: start_date
        required: true
        type: string
      - description: 结束日期 YYYY-MM-DD
        example: "2025-10-05"
        in: query
        name: end_date
        required: true
        type: string
      - default: daily
        description: 统计类型 daily/weekly/monthly
        enum:
        - daily
        - weekly
        - monthly
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.TrendData'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 访问趋势
      tags:
      - 统计管理
  /admin/stats/visits:
    get:
      consumes:
      - application/json
      description: 获取访问日志列表，支持分页查询
      parameters:
      - default: 1
        description: 页码
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
              type: object
      security:
      - BearerAuth: []
      summary: 访问日志
      tags:
      - 统计管理
...
  /collect:
    post:
      consumes:
      - application/json
      description: 前端埋点数据收集，记录页面访问等
      parameters:
      - description: 访问数据
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CollectRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: 数据收集
      tags:
      - 统计
...
securityDefinitions:
  BearerAuth:
    description: 在请求头中添加 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"